+++ greenfield_python_sdk/greenfield/object.py
@@ -2,11 +2,9 @@ import io
 import os
 from typing import Any, List, Tuple
 
-from betterproto import Casing
-
 from greenfield_python_sdk.blockchain_client import BlockchainClient
 from greenfield_python_sdk.key_manager import KeyManager
-from greenfield_python_sdk.models.bucket import VisibilityType
+from greenfield_python_sdk.models.bucket import EndPointOptions, VisibilityType
 from greenfield_python_sdk.models.eip712_messages.storage.object_url import (
     CANCEL_CREATE_OBJECT,
     CREATE_OBJECT,
@@ -17,8 +15,10 @@ from greenfield_python_sdk.models.eip712_messages.storage.policy_url import DELE
 from greenfield_python_sdk.models.object import (
     CreateObjectOptions,
     GetObjectOption,
+    ListObjectPoliciesOptions,
     ListObjectsOptions,
     ListObjectsResult,
+    ObjectMeta,
     PutObjectOptions,
 )
 from greenfield_python_sdk.models.request import Principal, PutPolicyOption, ResourceType
@@ -31,7 +31,6 @@ from greenfield_python_sdk.protos.greenfield.storage import (
     MsgPutPolicy,
     MsgUpdateObjectInfo,
     ObjectInfo,
-    QueryHeadBucketRequest,
     QueryHeadObjectByIdRequest,
     QueryHeadObjectRequest,
     QueryPolicyForAccountRequest,
@@ -106,7 +105,7 @@ class Object:
     async def get_object(self, bucket_name: str, object_name: str, opts: GetObjectOption) -> Tuple[Any, ObjectInfo]:
         sp = await self.blockchain_client.sp.get_first_in_service_storage_provider()
 
-        return await self.storage_client.object.get_object(bucket_name, object_name, opts, sp["operator_address"])
+        return await self.storage_client.object.get_object(bucket_name, object_name, sp["operator_address"], opts)
 
     async def get_object_head(self, bucket_name: str, object_name: str) -> ObjectInfo:
         object_info = await self.blockchain_client.storage.get_head_object(
@@ -203,7 +202,7 @@ class Object:
 
     async def list_objects(self, bucket_name: str, opts: ListObjectsOptions) -> ListObjectsResult:
         sp = await self.blockchain_client.sp.get_first_in_service_storage_provider()
-        return await self.storage_client.object.list_objects(bucket_name, opts, sp["operator_address"])
+        return await self.storage_client.object.list_objects(bucket_name, sp["operator_address"], opts)
 
     async def create_folder(self, bucket_name: str, object_name: str, opts: CreateObjectOptions) -> str:
         if object_name.endswith("/") == False:
@@ -233,3 +232,13 @@ class Object:
         _, data = await self.get_object(bucket_name, object_name, opts)
         file.write(data)
         file.close()
+
+    async def list_object_by_object_id(self, object_ids: List[int], opts: EndPointOptions) -> List[ObjectMeta]:
+        return await self.storage_client.object.list_object_by_object_id(object_ids, opts)
+
+    async def list_object_policies(
+        self, object_name: str, bucket_name: str, action_type: ActionType, opts: ListObjectPoliciesOptions
+    ):
+        # TODO: finish implementation
+        await self.storage_client.object.list_object_policies(object_name, bucket_name, action_type, opts)
+        raise NotImplementedError