+++ greenfield_python_sdk/models/eip712_messages/sp/msg_update_sp_status.py
@@ -1,0 +1,8 @@
+TYPES = {
+    "Msg1": [
+        {"name": "duration", "type": "string"},
+        {"name": "sp_address", "type": "string"},
+        {"name": "status", "type": "string"},
+        {"name": "type", "type": "string"},
+    ],
+}